requirements.txt
confluent-kafka==2.10.0
Django==5.2
djangorestframework==3.16.0
sqlparse==0.5.3
